Modified Dash with Movement
Modified the script and changed the logic, now I can control the distance and speed of the dash, while the equation is calculating the time of the dashing.
There is one thing left, to optimize the code and make the dash script work with the movement script interconnecting without disable the movement script.